# docs/swagger.yaml
definitions:
...
  domain.ZoneState:
    properties:
      comment:
        type: string
      created_at:
        type: string
      creator:
        type: string
      layout_id:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      state:
        type: string
      store_id:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
      zone_id:
        type: string
    type: object
  domain.ZoneStates:
    items:
      $ref: '#/definitions/domain.ZoneState'
    type: array
...
  infra.ChainZonesStatesResponse:
    properties:
      data:
        $ref: '#/definitions/domain.ZoneStates'
        type: object
      result_set:
        $ref: '#/definitions/infra.ResultSet'
        type: object
    type: object
...
  infra.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  infra.HTTPError:
    properties:
      message:
        type: object
    type: object
...
  infra.ResultSet:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: 1020@watcom.ru
    name: API Support
    url: https://helpdesk.watcom.ru
  description: |-
    This is a general service for interacting with layout configuration.
    Общий сервис для взаимодействия с конфигурацией схем размещения.
  license: {}
  title: LayoutConfig API
  version: 2.0/1.1.21
paths:
...
  /v2/chains/zones/states:
    get:
      description: |-
        get the zones states changes for zones with kind=service_channel, returns states online/offline
        layout_id (recommended parameter), if not pass datasource may be not correct
      parameters:
      - description: default=*
        in: query
        name: layout_id
        type: string
      - description: default=*
        in: query
        name: store_id
        type: string
      - description: default=*
        in: query
        name: zone_id
        type: string
      - description: default=0
        in: query
        name: offset
        type: integer
      - description: default=20
        in: query
        name: limit
        type: integer
      - description: ISO8601 datetime, default begin of day
        in: query
        name: from
        type: string
      - description: ISO8601 datetime, dafault current time
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.ChainZonesStatesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get zones (service_channel) states
      tags:
      - chains/zones
  /v2/chains/zones/states/attime:
    get:
      description: |-
        get the zones last states for zones with kind=service_channel, returns states online/offline
        layout_id (recommended parameter), if not pass datasource may be not correct
      parameters:
      - description: default=*
        in: query
        name: layout_id
        required: true
        type: string
      - description: uuid format
        in: query
        name: zone_id
        required: true
        type: string
      - description: ISO8601 datetime, default=current moment
        in: query
        name: at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ZoneState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get zones (service_channel) last states
      tags:
      - chains/zones
  /v2/chains/zones/states/last:
    get:
      description: |-
        get the zones last states for zones with kind=service_channel, returns states online/offline
        layout_id (recommended parameter), if not pass datasource may be not correct
      parameters:
      - description: default=*
        in: query
        name: layout_id
        type: string
      - description: default=*
        in: query
        name: store_id
        type: string
      - description: default=*
        in: query
        name: zone_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ZoneStates'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get zones (service_channel) last states
      tags:
      - chains/zones
...
swagger: "2.0"